fix: prevent redirect to step 1 after Excel import apply
- Add setIsScoresDirty(true) call in onConfirm callback of ImportPreviewModal
  to prevent useEffect sync from overwriting imported scores with empty data
- Invalidate existingGrades query cache on successful submitData mutation
  to ensure local state stays in sync with database after saving

+++ src/components/pages/bulk-grade-input/components/SettingsCard.tsx
@@ -142,7 +142,7 @@ export const SettingsCard: React.FC<SettingsCardProps> = ({
                         min={0}
                         max={100}
                         value={kkm}
-                        onChange={(e) => setKkm(Math.min(100, Math.max(0, parseInt(e.target.value) || 0)))}
+                        onChange={(e) => setKkm(Math.min(100, Math.max(0, parseFloat(e.target.value) || 0)))}
                         placeholder="75"
                     />
                 </div>
